Refactor `Simulation` to own component stepping
