+++ twine-core/src/transient/controllers.rs
@@ -1,0 +1,3 @@
+mod no_controller;
+
+pub use no_controller::NoController;
