+++ twine-core/src/transient/types.rs
@@ -2,12 +2,11 @@ mod time_increment;
 
 use std::{fmt::Debug, ops::Div};
 
-use thiserror::Error;
 use uom::si::f64::Time;
 
 use crate::Component;
 
-use super::{Controller, Integrator, Temporal};
+use super::Temporal;
 
 pub use time_increment::{TimeIncrement, TimeIncrementError};
 
@@ -39,33 +38,6 @@ where
     pub fn new(input: C::Input, output: C::Output) -> Self {
         Self { input, output }
     }
-}
-
-/// An error that can occur when advancing a simulation step using a controller.
-///
-/// This type aggregates failures from multiple sources involved in progressing
-/// the simulation:
-///
-/// - [`Integrator`]: Failed to generate a candidate input from history.
-/// - [`Controller`]: Failed to adjust the proposed input.
-/// - [`Component`]: Failed during evaluation on the final input.
-///
-/// It is returned by [`Controller::step`] and provides granular diagnostics for
-/// debugging simulation progression.
-#[derive(Debug, Error)]
-pub enum StepError<C, I, K>
-where
-    C: Component,
-    C::Input: Clone + Temporal,
-    I: Integrator<C>,
-    K: Controller<C>,
-{
-    #[error("Component failed: {0}")]
-    Component(C::Error),
-    #[error("Controller failed: {0}")]
-    Controller(K::Error),
-    #[error("Integrator failed: {0}")]
-    Integrator(I::Error),
 }
 
 /// The type representing the time derivative of `T`.
